ECN solutions

+++ src/dirs/sapientia-ecn.js
@@ -1,10 +1,12 @@
-import { existsSync, mkdirSync, readdirSync, writeFileSync } from 'fs';
+import { cpSync, existsSync, mkdirSync, readdirSync, rmSync, statSync, writeFileSync } from 'fs';
 import StreamZIP from 'node-stream-zip';
 
 const download = async s => {
     const name = 'InOut' + s.split('/')[1];
     const file = `inputs/${name}.zip`;
     const target = ['inputs', 'sapientia-ecn', ...s.split('.')[0].split('/').slice(1)];
+    target[target.length - 1] = target.at(-1)[0];
+    if (existsSync(target.join('/'))) return target;
     if (!existsSync(file)) writeFileSync(file, Buffer.from(
         await (await fetch(`https://ecn.ms.sapientia.ro/problems/${name}.zip`)).arrayBuffer(),
     ));
@@ -12,8 +14,17 @@ const download = async s => {
         .map((_, i) => target.slice(0, i + 1).join('/'))
         .forEach(s => existsSync(s) || mkdirSync(s));
     const zip = new StreamZIP.async({ file });
-    await zip.extract(name + '/' + target.at(-1), target.join('/'));
+    if (!await zip.extract(name + '/' + target.at(-1), target.join('/')))
+        await zip.extract(target.at(-1), target.join('/'));
     await zip.close();
+    for (const ent of readdirSync(target.join('/'))) {
+        const src = [...target, ent].join('/');
+        if (statSync(src).isDirectory()) {
+            for (const t of readdirSync(src))
+                cpSync(src + '/' + t, target.join('/') + '/' + ent.match(/\d+/)[0] + '.' + t.split('.').at(-1));
+            rmSync(src, { recursive: true });
+        }
+    }
     return target;
 };
